richacl: Move EVERYONE@ ACEs down the acl
The POSIX standard puts processes which are not the owner or a member in
the owning group or which match any ace other then EVERYONE@ on the
other file class.  We only know if a process is in the other class after
processing the entire acl.

Move all EVERYONE@ aces in the acl down in the acl so that at most a
single EVERYONE@ allow ace remains at the end.  Permissions which are
not explicitly allowed are implicitly denied, so EVERYONE@ deny ace is
needed.

The EVERYONE@ aces can be moved down the acl without changing the
permissions that the acl grants.  This transformation simplifies the
following algorithms, and eventually allows us to turn the final
EVERYONE@ allow ace into an entry for the other class.

Signed-off-by: Andreas Gruenbacher <agruen@kernel.org>
Signed-off-by: Aneesh Kumar K.V <aneesh.kumar@linux.vnet.ibm.com>