mm/debug_vm_pgtable/savedwrite: Enable savedwrite test with CONFIG_NUMA_BALANCING
Saved write support was added to track the write bit of a pte after
marking the pte protnone. This was done so that AUTONUMA can convert a write
pte to protnone and still track the old write bit. When converting it back we
set the pte write bit correctly thereby avoiding a write fault again. Hence
enable the test only when CONFIG_NUMA_BALANCING is enabled and use protnone
protflags.

Signed-off-by: Aneesh Kumar K.V <aneesh.kumar@linux.ibm.com>