powerpc/mm: Cleanup memory block size probing
Parse the device tree in early init to find the memory block size to be
used by the kernel. Consolidate the memory block size device tree parsing to
one helper and use that on both powernv and pseries. We still want to use
machine-specific callback because on all machine types other than powernv
and pseries we continue to return MIN_MEMORY_BLOCK_SIZE.

Signed-off-by: Aneesh Kumar K.V <aneesh.kumar@linux.ibm.com>